feat(homepage): add experiment api
